Refactored Natty Parser out into its own class - as a singleton to provide performance improvements (instantiation of it is slow)
